+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- Bug when loading menu on mobile
+
+### Changed
+
+- Changed Link paths in Submenu, fixed some translations
+
 ## [0.0.3] - 2023-11-07
 
 ### Added